Updated tests with new TM init access privilege
