+++ src/jmh/java/org/numenta/nupic/benchmarks/AbstractAlgorithmBenchmark.java
@@ -26,7 +26,7 @@ import org.numenta.nupic.Connections;
 import org.numenta.nupic.Parameters;
 import org.numenta.nupic.Parameters.KEY;
 import org.numenta.nupic.algorithms.SpatialPooler;
-import org.numenta.nupic.algorithms.OldTemporalMemory;
+import org.numenta.nupic.algorithms.TemporalMemory;
 import org.numenta.nupic.encoders.ScalarEncoder;
 import org.openjdk.jmh.annotations.Scope;
 import org.openjdk.jmh.annotations.Setup;
@@ -38,7 +38,7 @@ public abstract class AbstractAlgorithmBenchmark {
     protected int[] SDR;
     protected ScalarEncoder encoder;
     protected SpatialPooler pooler;
-    protected OldTemporalMemory temporalMemory;
+    protected TemporalMemory temporalMemory;
     protected Connections memory;
 
     @Setup
@@ -65,8 +65,8 @@ public abstract class AbstractAlgorithmBenchmark {
         pooler = new SpatialPooler();
         pooler.init(memory);
 
-        temporalMemory = new OldTemporalMemory();
-        temporalMemory.init(memory);
+        temporalMemory = new TemporalMemory();
+        TemporalMemory.init(memory);
     }
 
     /**
