+++ src/main/java/org/numenta/nupic/ComputeCycle.java
@@ -31,7 +31,7 @@ import java.util.Optional;
 import java.util.Set;
 
 import org.numenta.nupic.Connections.SegmentOverlap;
-import org.numenta.nupic.algorithms.OldTemporalMemory;
+import org.numenta.nupic.algorithms.TemporalMemory;
 import org.numenta.nupic.model.Cell;
 import org.numenta.nupic.model.Column;
 import org.numenta.nupic.model.DistalDendrite;
@@ -39,7 +39,7 @@ import org.numenta.nupic.util.Tuple;
 
 /**
  * Contains a snapshot of the state attained during one computational
- * call to the {@link OldTemporalMemory}. The {@code TemporalMemory} uses
+ * call to the {@link TemporalMemory}. The {@code TemporalMemory} uses
  * data from previous compute cycles to derive new data for the current cycle
  * through a comparison between states of those different cycles, therefore
  * this state container is necessary.
@@ -76,7 +76,7 @@ public class ComputeCycle implements Persistable {
     /**
      * Constructs a new {@code ComputeCycle} initialized with
      * the connections relevant to the current calling {@link Thread} for
-     * the specified {@link OldTemporalMemory}
+     * the specified {@link TemporalMemory}
      * 
      * @param   c       the current connections state of the TemporalMemory
      */
@@ -156,7 +156,7 @@ public class ComputeCycle implements Persistable {
     
     /**
      * Returns the Set of matching {@link DistalDendrite}s from 
-     * {@link OldTemporalMemory#computePredictiveCells(Connections, ComputeCycle, Map)}
+     * {@link TemporalMemory#computePredictiveCells(Connections, ComputeCycle, Map)}
      * @return
      */
     public Set<DistalDendrite> matchingSegments() {
@@ -165,7 +165,7 @@ public class ComputeCycle implements Persistable {
     
     /**
      * Returns the Set of matching {@link Cell}s from
-     * {@link OldTemporalMemory#computePredictiveCells(Connections, ComputeCycle, Map)}
+     * {@link TemporalMemory#computePredictiveCells(Connections, ComputeCycle, Map)}
      * @return
      */
     public Set<Cell> matchingCells() {
