+++ src/main/java/org/numenta/nupic/algorithms/TemporalMemory.java
@@ -38,7 +38,7 @@ public class TemporalMemory implements ComputeDecorator, Serializable {
      * anatomy needed by this {@code TemporalMemory} to implement its algorithms.
      * 
      * The connections object holds the {@link Column} and {@link Cell} infrastructure,
-     * and is used by both the {@link SpatialPooler} and {@link OldTemporalMemory}. Either of
+     * and is used by both the {@link SpatialPooler} and {@link TemporalMemory}. Either of
      * these can be used separately, and therefore this Connections object may have its
      * Columns and Cells initialized by either the init method of the SpatialPooler or the
      * init method of the TemporalMemory. We check for this so that complete initialization
