+++ src/main/java/org/numenta/nupic/monitor/mixin/MonitorMixinBase.java
@@ -47,7 +47,7 @@ import com.bethecoder.table.spec.AsciiTable;
  *      <li>Create MonitorMixinBase extension mixin type (interface) (this does most of the mixin work) see {@link TemporalMemoryMonitorMixin}</li>
  *      <li>Create Extension of both {@link MonitorMixinBase} and interface (see {@link MonitoredTemporalMemory})</li>
  *      <li>Create constructor in extension class which takes above interface (A) which is the target class which must implement (A)</li>
- *      <li>Make the Original class (see {@link OldTemporalMemory}) to be tested, implement the Decorator interface (A)</li>
+ *      <li>Make the Original class (see {@link TemporalMemory}) to be tested, implement the Decorator interface (A)</li>
  *      <li>Use the "joining" extension class as you would the original class</li>
  * </OL>
  *
