+++ src/main/java/org/numenta/nupic/network/Inference.java
@@ -29,7 +29,7 @@ import org.numenta.nupic.Persistable;
 import org.numenta.nupic.algorithms.CLAClassifier;
 import org.numenta.nupic.algorithms.Classification;
 import org.numenta.nupic.algorithms.SpatialPooler;
-import org.numenta.nupic.algorithms.OldTemporalMemory;
+import org.numenta.nupic.algorithms.TemporalMemory;
 import org.numenta.nupic.encoders.Encoder;
 import org.numenta.nupic.model.Cell;
 import org.numenta.nupic.util.NamedTuple;
@@ -125,7 +125,7 @@ public interface Inference extends Persistable {
      */
     public int[] getFeedForwardSparseActives();
     /**
-     * Returns the column activation from a {@link OldTemporalMemory}
+     * Returns the column activation from a {@link TemporalMemory}
      * @return
      */
     public Set<Cell> getActiveCells();
