+++ src/main/java/org/numenta/nupic/network/Layer.java
@@ -44,7 +44,7 @@ import org.numenta.nupic.algorithms.Anomaly;
 import org.numenta.nupic.algorithms.CLAClassifier;
 import org.numenta.nupic.algorithms.Classification;
 import org.numenta.nupic.algorithms.SpatialPooler;
-import org.numenta.nupic.algorithms.OldTemporalMemory;
+import org.numenta.nupic.algorithms.TemporalMemory;
 import org.numenta.nupic.encoders.DateEncoder;
 import org.numenta.nupic.encoders.Encoder;
 import org.numenta.nupic.encoders.EncoderTuple;
@@ -186,7 +186,7 @@ public class Layer<T> implements Persistable {
     protected HTMSensor<?> sensor;
     protected MultiEncoder encoder;
     protected SpatialPooler spatialPooler;
-    protected OldTemporalMemory temporalMemory;
+    protected TemporalMemory temporalMemory;
     private Boolean autoCreateClassifiers;
     private Anomaly anomalyComputer;
 
@@ -203,11 +203,11 @@ public class Layer<T> implements Persistable {
     protected int[] feedForwardActiveColumns;
     /** Active column indexes from the {@link SpatialPooler} at time "t" */
     private int[] feedForwardSparseActives;
-    /** Predictive {@link Cell}s in the {@link OldTemporalMemory} at time "t - 1" */
+    /** Predictive {@link Cell}s in the {@link TemporalMemory} at time "t - 1" */
     private Set<Cell> previousPredictiveCells;
-    /** Predictive {@link Cell}s in the {@link OldTemporalMemory} at time "t" */
+    /** Predictive {@link Cell}s in the {@link TemporalMemory} at time "t" */
     private Set<Cell> predictiveCells;
-    /** Active {@link Cell}s in the {@link OldTemporalMemory} at time "t" */
+    /** Active {@link Cell}s in the {@link TemporalMemory} at time "t" */
     private Set<Cell> activeCells;
 
     /** Used to track and document the # of records processed */
@@ -363,12 +363,12 @@ public class Layer<T> implements Persistable {
      * @param e                         (optional) The Network API only uses a {@link MultiEncoder} at
      *                                  the top level because of its ability to delegate to child encoders.
      * @param sp                        (optional) {@link SpatialPooler}
-     * @param tm                        (optional) {@link OldTemporalMemory}
+     * @param tm                        (optional) {@link TemporalMemory}
      * @param autoCreateClassifiers     (optional) Indicates that the {@link Parameters} object
      *                                  contains the configurations necessary to create the required encoders.
      * @param a                         (optional) An {@link Anomaly} computer.
      */
-    public Layer(Parameters params, MultiEncoder e, SpatialPooler sp, OldTemporalMemory tm, Boolean autoCreateClassifiers, Anomaly a) {
+    public Layer(Parameters params, MultiEncoder e, SpatialPooler sp, TemporalMemory tm, Boolean autoCreateClassifiers, Anomaly a) {
 
         // Make sure we have a valid parameters object
         if(params == null) {
@@ -535,7 +535,7 @@ public class Layer<T> implements Persistable {
 
         // Let the TemporalMemory initialize the matrix with its requirements
         if(temporalMemory != null) {
-            temporalMemory.init(connections);
+            TemporalMemory.init(connections);
         }
 
         this.feedForwardActiveColumns = new int[connections.getNumColumns()];
@@ -549,7 +549,7 @@ public class Layer<T> implements Persistable {
     
     /**
      * Called from {@link FunctionFactory#createSpatialFunc(SpatialPooler)} and from {@link #close()}
-     * to calculate the size of the input vector given the output source either being a {@link OldTemporalMemory}
+     * to calculate the size of the input vector given the output source either being a {@link TemporalMemory}
      * or a {@link SpatialPooler} - from this {@link Region} or a previous {@link Region}.
      * 
      * @return  the length of the input vector
@@ -590,7 +590,7 @@ public class Layer<T> implements Persistable {
 
     /**
      * For internal use only. Returns a flag indicating whether this {@link Layer}
-     * contains a {@link OldTemporalMemory}
+     * contains a {@link TemporalMemory}
      * @return
      */
     boolean hasTM() {
@@ -804,12 +804,12 @@ public class Layer<T> implements Persistable {
     }
 
     /**
-     * Adds a {@link OldTemporalMemory} to this {@code Layer}
+     * Adds a {@link TemporalMemory} to this {@code Layer}
      * 
      * @param tm    the added TemporalMemory
      * @return this Layer instance (in fluent-style)
      */
-    public Layer<T> add(OldTemporalMemory tm) {
+    public Layer<T> add(TemporalMemory tm) {
         if(isClosed) {
             throw new IllegalStateException("Layer already \"closed\"");
         }
@@ -1194,7 +1194,7 @@ public class Layer<T> implements Persistable {
     }
 
     /**
-     * Returns the {@link Cell}s activated in the {@link OldTemporalMemory} at time
+     * Returns the {@link Cell}s activated in the {@link TemporalMemory} at time
      * "t"
      * 
      * @return
@@ -1263,7 +1263,7 @@ public class Layer<T> implements Persistable {
     }
 
     /**
-     * Resets the {@link OldTemporalMemory} if it exists.
+     * Resets the {@link TemporalMemory} if it exists.
      */
     public void reset() {
         if(temporalMemory == null) {
@@ -1275,7 +1275,7 @@ public class Layer<T> implements Persistable {
 
     /**
      * Returns a flag indicating whether this {@code Layer} contains a
-     * {@link OldTemporalMemory}.
+     * {@link TemporalMemory}.
      * 
      * @return
      */
@@ -1784,7 +1784,7 @@ public class Layer<T> implements Persistable {
                 } else {
                     o = o.map(factory.createSpatialFunc(spatialPooler));
                 }
-            } else if(node instanceof OldTemporalMemory) {
+            } else if(node instanceof TemporalMemory) {
                 o = o.map(factory.createTemporalFunc(temporalMemory));
             }
         }
@@ -1947,7 +1947,7 @@ public class Layer<T> implements Persistable {
     }
 
     /**
-     * Called internally to invoke the {@link OldTemporalMemory}
+     * Called internally to invoke the {@link TemporalMemory}
      * 
      * @param input     the current input vector
      * @param mi        the current input inference container
@@ -2293,7 +2293,7 @@ public class Layer<T> implements Persistable {
             };
         }
 
-        public Func1<ManualInput, ManualInput> createTemporalFunc(final OldTemporalMemory tm) {
+        public Func1<ManualInput, ManualInput> createTemporalFunc(final TemporalMemory tm) {
             return new Func1<ManualInput, ManualInput>() {
 
                 @Override
