+++ src/test/java/org/numenta/nupic/integration/AbstractTemporalMemoryTest.java
@@ -29,7 +29,7 @@ import java.util.Set;
 import org.numenta.nupic.Connections;
 import org.numenta.nupic.Parameters;
 import org.numenta.nupic.Parameters.KEY;
-import org.numenta.nupic.algorithms.OldTemporalMemory;
+import org.numenta.nupic.algorithms.TemporalMemory;
 import org.numenta.nupic.datagen.PatternMachine;
 import org.numenta.nupic.datagen.SequenceMachine;
 import org.numenta.nupic.monitor.MonitoredTemporalMemory;
@@ -37,7 +37,7 @@ import org.numenta.nupic.util.ArrayUtils;
 
 
 public class AbstractTemporalMemoryTest {
-    protected OldTemporalMemory temporalMemory;
+    protected TemporalMemory temporalMemory;
     protected Parameters parameters;
     protected Connections connections;
     protected PatternMachine patternMachine;
@@ -50,8 +50,8 @@ public class AbstractTemporalMemoryTest {
         this.connections = new Connections();
         parameters.apply(connections);
         
-        temporalMemory = new OldTemporalMemory();
-        temporalMemory.init(connections);
+        temporalMemory = new TemporalMemory();
+        TemporalMemory.init(connections);
         tm = new MonitoredTemporalMemory(temporalMemory, connections);
         
         this.patternMachine = pm;
