+++ src/test/java/org/numenta/nupic/network/LayerTest.java
@@ -45,7 +45,7 @@ import org.numenta.nupic.algorithms.Anomaly;
 import org.numenta.nupic.algorithms.Anomaly.Mode;
 import org.numenta.nupic.algorithms.CLAClassifier;
 import org.numenta.nupic.algorithms.SpatialPooler;
-import org.numenta.nupic.algorithms.OldTemporalMemory;
+import org.numenta.nupic.algorithms.TemporalMemory;
 import org.numenta.nupic.datagen.ResourceLocator;
 import org.numenta.nupic.encoders.MultiEncoder;
 import org.numenta.nupic.network.Layer.FunctionFactory;
@@ -118,7 +118,7 @@ public class LayerTest extends ObservableTestBase {
         p.set(KEY.RANDOM, new UniversalRandom(42));
 
         MultiEncoder me = MultiEncoder.builder().name("").build();
-        Layer<Map<String, Object>> l = new Layer<>(p, me, new SpatialPooler(), new OldTemporalMemory(), Boolean.TRUE, null);
+        Layer<Map<String, Object>> l = new Layer<>(p, me, new SpatialPooler(), new TemporalMemory(), Boolean.TRUE, null);
 
         // Test that we get the expected exception if there hasn't been any processing.
         try {
@@ -155,7 +155,7 @@ public class LayerTest extends ObservableTestBase {
     @Test
     public void testResetMethod() {
         Parameters p = NetworkTestHarness.getParameters().copy();
-        Layer<?> l = Network.createLayer("l1", p).add(new OldTemporalMemory());
+        Layer<?> l = Network.createLayer("l1", p).add(new TemporalMemory());
         try {
             l.reset();
             assertTrue(l.hasTemporalMemory());
@@ -176,7 +176,7 @@ public class LayerTest extends ObservableTestBase {
     public void testResetRecordNum() {
         Parameters p = NetworkTestHarness.getParameters().copy();
         @SuppressWarnings("unchecked")
-        Layer<int[]> l = (Layer<int[]>)Network.createLayer("l1", p).add(new OldTemporalMemory());
+        Layer<int[]> l = (Layer<int[]>)Network.createLayer("l1", p).add(new TemporalMemory());
         l.subscribe(new Observer<Inference>() {
             @Override public void onCompleted() {}
             @Override public void onError(Throwable e) { e.printStackTrace(); }
@@ -818,7 +818,7 @@ public class LayerTest extends ObservableTestBase {
         final int[][] inputs = { input1, input2, input3, input4, input5, input6, input7 };
 
         
-        Layer<int[]> l = new Layer<>(p, null, null, new OldTemporalMemory(), null, null);
+        Layer<int[]> l = new Layer<>(p, null, null, new TemporalMemory(), null, null);
         
         int timeUntilStable = 600;
 
@@ -872,7 +872,7 @@ public class LayerTest extends ObservableTestBase {
         inputs[5] = new int[] { 0, 0, 0, 0, 1, 1, 1, 0 };
         inputs[6] = new int[] { 0, 0, 0, 0, 0, 1, 1, 1 };
 
-        Layer<int[]> l = new Layer<>(p, null, new SpatialPooler(), new OldTemporalMemory(), null, null);
+        Layer<int[]> l = new Layer<>(p, null, new SpatialPooler(), new TemporalMemory(), null, null);
         TestObserver<Inference> tester;
         l.subscribe(tester = new TestObserver<Inference>() {
             @Override public void onCompleted() {}
@@ -970,7 +970,7 @@ public class LayerTest extends ObservableTestBase {
         p.set(KEY.RANDOM, new UniversalRandom(42));
 
         MultiEncoder me = MultiEncoder.builder().name("").build();
-        Layer<Map<String, Object>> l = new Layer<>(p, me, new SpatialPooler(), new OldTemporalMemory(), Boolean.TRUE, null);
+        Layer<Map<String, Object>> l = new Layer<>(p, me, new SpatialPooler(), new TemporalMemory(), Boolean.TRUE, null);
         TestObserver<Inference> tester;
         l.subscribe(tester = new TestObserver<Inference>() {
             @Override public void onCompleted() {}
@@ -1010,7 +1010,7 @@ public class LayerTest extends ObservableTestBase {
         p.set(KEY.SP_PRIMER_DELAY, PRIME_COUNT);
 
         MultiEncoder me = MultiEncoder.builder().name("").build();
-        Layer<Map<String, Object>> l = new Layer<>(p, me, new SpatialPooler(), new OldTemporalMemory(), Boolean.TRUE, null);
+        Layer<Map<String, Object>> l = new Layer<>(p, me, new SpatialPooler(), new TemporalMemory(), Boolean.TRUE, null);
         TestObserver<Inference> tester;
         l.subscribe(tester = new TestObserver<Inference>() {
             @Override public void onCompleted() {}
@@ -1057,7 +1057,7 @@ public class LayerTest extends ObservableTestBase {
         p.set(KEY.SP_PRIMER_DELAY, PRIME_COUNT);
 
         MultiEncoder me = MultiEncoder.builder().name("").build();
-        Layer<Map<String, Object>> l = new Layer<>(p, me, new SpatialPooler(), new OldTemporalMemory(), Boolean.TRUE, null);
+        Layer<Map<String, Object>> l = new Layer<>(p, me, new SpatialPooler(), new TemporalMemory(), Boolean.TRUE, null);
 
         int[][] inputs = new int[7][8];
         inputs[0] = new int[] { 1, 1, 0, 0, 0, 0, 0, 1 };
@@ -1136,7 +1136,7 @@ public class LayerTest extends ObservableTestBase {
         assertTrue(cellsPerColumn > 0);
 
         MultiEncoder me = MultiEncoder.builder().name("").build();
-        final Layer<Map<String, Object>> l = new Layer<>(p, me, new SpatialPooler(), new OldTemporalMemory(), Boolean.TRUE, null);
+        final Layer<Map<String, Object>> l = new Layer<>(p, me, new SpatialPooler(), new TemporalMemory(), Boolean.TRUE, null);
 
         TestObserver<Inference> tester;
         l.subscribe(tester = new TestObserver<Inference>() {
@@ -1204,7 +1204,7 @@ public class LayerTest extends ObservableTestBase {
         p.set(KEY.RANDOM, new MersenneTwister(42));
 
         MultiEncoder me = MultiEncoder.builder().name("").build();
-        final Layer<Map<String, Object>> l = new Layer<>(p, me, new SpatialPooler(), new OldTemporalMemory(), Boolean.TRUE, null);
+        final Layer<Map<String, Object>> l = new Layer<>(p, me, new SpatialPooler(), new TemporalMemory(), Boolean.TRUE, null);
 
         final List<int[]> emissions = new ArrayList<int[]>();
         Observable<Inference> o = l.observe();
@@ -1267,7 +1267,7 @@ public class LayerTest extends ObservableTestBase {
         Layer<?> l = Network.createLayer("TestLayer", p)
             .alterParameter(KEY.AUTO_CLASSIFY, true)
             .add(anomalyComputer)
-            .add(new OldTemporalMemory())
+            .add(new TemporalMemory())
             .add(new SpatialPooler())
             .add(Sensor.create(
                 FileSensor::create, 
@@ -1319,7 +1319,7 @@ public class LayerTest extends ObservableTestBase {
         Layer<?> l = Network.createLayer("TestLayer", p)
             .alterParameter(KEY.AUTO_CLASSIFY, true)
             .add(anomalyComputer)
-            .add(new OldTemporalMemory())
+            .add(new TemporalMemory())
             .add(new SpatialPooler())
             .add(Sensor.create(
                 FileSensor::create, 
@@ -1373,7 +1373,7 @@ public class LayerTest extends ObservableTestBase {
         l = Network.createLayer("TestLayer", p)
             .alterParameter(KEY.AUTO_CLASSIFY, true)
             .add(anomalyComputer)
-            .add(new OldTemporalMemory())
+            .add(new TemporalMemory())
             .add(new SpatialPooler())
             .add(anomalyComputer)
             .add(MultiEncoder.builder().name("").build());
@@ -1452,8 +1452,8 @@ public class LayerTest extends ObservableTestBase {
     @Test
     public void testEquality() {
         Parameters p = Parameters.getAllDefaultParameters();
-        Layer<Map<String, Object>> l = new Layer<>(p, null, new SpatialPooler(), new OldTemporalMemory(), Boolean.TRUE, null);
-        Layer<Map<String, Object>> l2 = new Layer<>(p, null, new SpatialPooler(), new OldTemporalMemory(), Boolean.TRUE, null);
+        Layer<Map<String, Object>> l = new Layer<>(p, null, new SpatialPooler(), new TemporalMemory(), Boolean.TRUE, null);
+        Layer<Map<String, Object>> l2 = new Layer<>(p, null, new SpatialPooler(), new TemporalMemory(), Boolean.TRUE, null);
         
         assertTrue(l.equals(l));
         assertFalse(l.equals(null));
@@ -1481,7 +1481,7 @@ public class LayerTest extends ObservableTestBase {
     @Test
     public void testInferInputDimensions() {
         Parameters p = Parameters.getAllDefaultParameters();
-        Layer<Map<String, Object>> l = new Layer<>(p, null, new SpatialPooler(), new OldTemporalMemory(), Boolean.TRUE, null);
+        Layer<Map<String, Object>> l = new Layer<>(p, null, new SpatialPooler(), new TemporalMemory(), Boolean.TRUE, null);
         
         int[] dims = l.inferInputDimensions(16384, 2);
         assertTrue(Arrays.equals(new int[] { 128, 128 }, dims));
