+++ src/test/java/org/numenta/nupic/network/NetworkTest.java
@@ -44,7 +44,7 @@ import org.numenta.nupic.Parameters.KEY;
 import org.numenta.nupic.algorithms.Anomaly;
 import org.numenta.nupic.algorithms.Anomaly.Mode;
 import org.numenta.nupic.algorithms.SpatialPooler;
-import org.numenta.nupic.algorithms.OldTemporalMemory;
+import org.numenta.nupic.algorithms.TemporalMemory;
 import org.numenta.nupic.datagen.ResourceLocator;
 import org.numenta.nupic.encoders.MultiEncoder;
 import org.numenta.nupic.network.sensor.FileSensor;
@@ -81,7 +81,7 @@ public class NetworkTest extends ObservableTestBase {
         Parameters p = NetworkTestHarness.getParameters();
         Network network = new Network("ResetTestNetwork", p)
             .add(Network.createRegion("r1")
-                .add(Network.createLayer("l1", p).add(new OldTemporalMemory())));
+                .add(Network.createLayer("l1", p).add(new TemporalMemory())));
         try {
             network.reset();
             assertTrue(network.lookup("r1").lookup("l1").hasTemporalMemory());
@@ -105,7 +105,7 @@ public class NetworkTest extends ObservableTestBase {
         Parameters p = NetworkTestHarness.getParameters();
         Network network = new Network("ResetRecordNumNetwork", p)
             .add(Network.createRegion("r1")
-                .add(Network.createLayer("l1", p).add(new OldTemporalMemory())));
+                .add(Network.createLayer("l1", p).add(new TemporalMemory())));
         network.observe().subscribe(new Observer<Inference>() {
             @Override public void onCompleted() {}
             @Override public void onError(Throwable e) { e.printStackTrace(); }
@@ -216,7 +216,7 @@ public class NetworkTest extends ObservableTestBase {
                 .add(Network.createLayer("1", p)
                     .alterParameter(KEY.AUTO_CLASSIFY, Boolean.TRUE)
                     .add(Anomaly.create())
-                    .add(new OldTemporalMemory())
+                    .add(new TemporalMemory())
                     .add(new SpatialPooler())
                     .add(Sensor.create(FileSensor::create, SensorParams.create(
                         Keys::path, "", ResourceLocator.path("rec-center-hourly.csv"))))));
@@ -277,7 +277,7 @@ public class NetworkTest extends ObservableTestBase {
                 .add(Network.createLayer("1", p)
                     .alterParameter(KEY.AUTO_CLASSIFY, Boolean.TRUE)
                     .add(Anomaly.create())
-                    .add(new OldTemporalMemory())
+                    .add(new TemporalMemory())
                     .add(new SpatialPooler())
                     .add(Sensor.create(FileSensor::create, SensorParams.create(
                         Keys::path, "", ResourceLocator.path("rec-center-hourly.csv"))))));
@@ -493,7 +493,7 @@ public class NetworkTest extends ObservableTestBase {
                 .add(Network.createLayer("1", p)
                     .alterParameter(KEY.AUTO_CLASSIFY, Boolean.TRUE)
                     .add(Anomaly.create())
-                    .add(new OldTemporalMemory())
+                    .add(new TemporalMemory())
                     .add(new SpatialPooler())
                     .add(Sensor.create(FileSensor::create, SensorParams.create(
                         Keys::path, "", ResourceLocator.path("rec-center-hourly.csv"))))));
@@ -555,12 +555,12 @@ public class NetworkTest extends ObservableTestBase {
             .add(Network.createRegion("r1")
                 .add(Network.createLayer("2", p)
                     .add(Anomaly.create())
-                    .add(new OldTemporalMemory())
+                    .add(new TemporalMemory())
                     .add(new SpatialPooler())))
             .add(Network.createRegion("r2")
                 .add(Network.createLayer("1", p)
                     .alterParameter(KEY.AUTO_CLASSIFY, Boolean.TRUE)
-                    .add(new OldTemporalMemory())
+                    .add(new TemporalMemory())
                     .add(new SpatialPooler())
                     .add(Sensor.create(FileSensor::create, SensorParams.create(
                         Keys::path, "", ResourceLocator.path("rec-center-hourly.csv"))))))
@@ -642,26 +642,26 @@ public class NetworkTest extends ObservableTestBase {
                         .add(Sensor.create(FileSensor::create, SensorParams.create(
                             Keys::path, "", ResourceLocator.path("rec-center-hourly.csv"))))
                         .add(new SpatialPooler())
-                        .add(new OldTemporalMemory())
+                        .add(new TemporalMemory())
                         .add(Anomaly.create(anomalyParams))
                 )
                     .add(Network.createLayer("1", p)            // Add another Layer, and the Region internally connects it to the 
                         .add(new SpatialPooler())               // previously added Layer
                         .using(new Connections())               // Test adding connections after one element and before another
-                        .add(new OldTemporalMemory())
+                        .add(new TemporalMemory())
                         .add(Anomaly.create(anomalyParams))
                 ))            
                 .add(Network.createRegion("r2")
                     .add(Network.createLayer("2/3", p)
                         .add(new SpatialPooler())
                         .using(new Connections()) // Test adding connections after one element and before another
-                        .add(new OldTemporalMemory())
+                        .add(new TemporalMemory())
                         .add(Anomaly.create(anomalyParams))
                 ))
                 .add(Network.createRegion("r3")
                     .add(Network.createLayer("1", p)
                         .add(new SpatialPooler())
-                        .add(new OldTemporalMemory())
+                        .add(new TemporalMemory())
                         .add(Anomaly.create(anomalyParams))
                             .using(new Connections()) // Test adding connections after elements which use them.
                 ))
@@ -697,7 +697,7 @@ public class NetworkTest extends ObservableTestBase {
                 .add(Network.createLayer("2", p)
                     .add(Anomaly.create(params)))
                 .add(Network.createLayer("3", p)
-                    .add(new OldTemporalMemory()))
+                    .add(new TemporalMemory()))
                 .add(Network.createLayer("4", p)
                     .add(new SpatialPooler())
                     .add(MultiEncoder.builder().name("").build()))
@@ -767,7 +767,7 @@ public class NetworkTest extends ObservableTestBase {
             .add(Network.createRegion("r1")
                 .add(Network.createLayer("1", p)
                     .alterParameter(KEY.AUTO_CLASSIFY, Boolean.TRUE)
-                    .add(new OldTemporalMemory())
+                    .add(new TemporalMemory())
                     .add(new SpatialPooler())
                     .add(MultiEncoder.builder().name("").build())));
         
@@ -818,7 +818,7 @@ public class NetworkTest extends ObservableTestBase {
                 .add(Network.createLayer("2", p)
                     .add(Anomaly.create(params)))
                 .add(Network.createLayer("3", p)
-                    .add(new OldTemporalMemory()))
+                    .add(new TemporalMemory()))
                 .add(Network.createLayer("4", p)
                     .add(new SpatialPooler())
                     .add(MultiEncoder.builder().name("").build()))
@@ -842,7 +842,7 @@ public class NetworkTest extends ObservableTestBase {
                 .add(Network.createLayer("2", p)
                     .add(Anomaly.create(params)))
                 .add(Network.createLayer("3", p)
-                    .add(new OldTemporalMemory()))
+                    .add(new TemporalMemory()))
                 .add(Network.createLayer("4", p)
                     .add(new SpatialPooler())
                     .add(Sensor.create(FileSensor::create, SensorParams.create(
@@ -861,7 +861,7 @@ public class NetworkTest extends ObservableTestBase {
                 .add(Network.createRegion("r1")
                     .add(Network.createLayer("1", p)
                         .alterParameter(KEY.AUTO_CLASSIFY, Boolean.TRUE)
-                        .add(new OldTemporalMemory())
+                        .add(new TemporalMemory())
                         .add(new SpatialPooler())
                         .add(Sensor.create(FileSensor::create, SensorParams.create(
                             Keys::path, "", ResourceLocator.path("rec-center-hourly.csv"))))));
@@ -899,7 +899,7 @@ public class NetworkTest extends ObservableTestBase {
             .add(Network.createRegion("r1")
                 .add(Network.createLayer("1", p)
                     .add(Anomaly.create())
-                    .add(new OldTemporalMemory())
+                    .add(new TemporalMemory())
                     .add(new SpatialPooler())
                     .add(htmSensor)));
 
@@ -961,7 +961,7 @@ public class NetworkTest extends ObservableTestBase {
                 .add(Network.createLayer("1", p)
                     .alterParameter(KEY.AUTO_CLASSIFY, Boolean.TRUE)
                     .add(Anomaly.create())
-                    .add(new OldTemporalMemory())
+                    .add(new TemporalMemory())
                     .add(new SpatialPooler())
                     .add(htmSensor)));
 
@@ -1020,12 +1020,12 @@ public class NetworkTest extends ObservableTestBase {
             .add(Network.createRegion("r1")
                 .add(Network.createLayer("2", p)
                     .add(Anomaly.create())
-                    .add(new OldTemporalMemory())
+                    .add(new TemporalMemory())
                     .add(new SpatialPooler())))
             .add(Network.createRegion("r2")
                 .add(Network.createLayer("1", p)
                     .alterParameter(KEY.AUTO_CLASSIFY, Boolean.TRUE)
-                    .add(new OldTemporalMemory())
+                    .add(new TemporalMemory())
                     .add(new SpatialPooler())
                     .add(Sensor.create(FileSensor::create, SensorParams.create(
                         Keys::path, "", ResourceLocator.path("rec-center-hourly.csv"))))))
@@ -1048,7 +1048,7 @@ public class NetworkTest extends ObservableTestBase {
             .add(Network.createRegion("r1")
                 .add(Network.createLayer("2", p)
                     .add(Anomaly.create())
-                    .add(new OldTemporalMemory())
+                    .add(new TemporalMemory())
                     .add(new SpatialPooler())))
             .add(Network.createRegion("r2")
                 .add(Network.createLayer("1", p)
@@ -1076,7 +1076,7 @@ public class NetworkTest extends ObservableTestBase {
             .add(Network.createRegion("r1")
                     .add(Network.createLayer("2", p)
                             .add(Anomaly.create())
-                            .add(new OldTemporalMemory())
+                            .add(new TemporalMemory())
                                     //.add(new SpatialPooler())
                             .close()));
         
@@ -1097,7 +1097,7 @@ public class NetworkTest extends ObservableTestBase {
                 .add(Network.createRegion("r1")
                         .add(Network.createLayer("2", p)
                                 .add(Anomaly.create())
-                                .add(new OldTemporalMemory())
+                                .add(new TemporalMemory())
                                         .add(new SpatialPooler())
                                 .close()));
 
@@ -1141,7 +1141,7 @@ public class NetworkTest extends ObservableTestBase {
                     .add(new SpatialPooler()))
                 .add(Network.createLayer("2", p)
                     .add(Anomaly.create())
-                    .add(new OldTemporalMemory())
+                    .add(new TemporalMemory())
                     .add(new SpatialPooler()))
                 .connect("1", "2"));
                     
@@ -1255,7 +1255,7 @@ public class NetworkTest extends ObservableTestBase {
             .add(Network.createLayer("1", p)
                 .alterParameter(KEY.AUTO_CLASSIFY, true)
                 .add(Anomaly.create())
-                .add(new OldTemporalMemory())
+                .add(new TemporalMemory())
                 .add(new SpatialPooler())
                 .add(sensor)));
