+++ src/test/java/org/numenta/nupic/network/PersistenceAPITest.java
@@ -63,7 +63,7 @@ import org.numenta.nupic.algorithms.CLAClassifier;
 import org.numenta.nupic.algorithms.Classification;
 import org.numenta.nupic.algorithms.Sample;
 import org.numenta.nupic.algorithms.SpatialPooler;
-import org.numenta.nupic.algorithms.OldTemporalMemory;
+import org.numenta.nupic.algorithms.TemporalMemory;
 import org.numenta.nupic.datagen.ResourceLocator;
 import org.numenta.nupic.encoders.DateEncoder;
 import org.numenta.nupic.encoders.MultiEncoder;
@@ -137,7 +137,7 @@ public class PersistenceAPITest extends ObservableTestBase {
         Network network = Network.create("test network", p).add(Network.createRegion("r1")
             .add(Network.createLayer("1", p)
                 .add(Anomaly.create())
-                .add(new OldTemporalMemory())
+                .add(new TemporalMemory())
                 .add(new SpatialPooler())));
         
         PersistenceAPI pa = Persistence.get(new SerialConfig(null, SerialConfig.SERIAL_TEST_DIR));
@@ -198,8 +198,7 @@ public class PersistenceAPITest extends ObservableTestBase {
         Connections con = new Connections();
         p.apply(con);
 
-        OldTemporalMemory tm = new OldTemporalMemory();
-        tm.init(con);
+        TemporalMemory.init(con);
 
         SerialConfig config = new SerialConfig("testSerializeConnections", SerialConfig.SERIAL_TEST_DIR);
         PersistenceAPI api = Persistence.get(config);
@@ -1690,7 +1689,7 @@ public class PersistenceAPITest extends ObservableTestBase {
             .add(Network.createRegion("r1")
                 .add(l2 = Network.createLayer("2", p)
                     .add(Anomaly.create())
-                    .add(new OldTemporalMemory()))
+                    .add(new TemporalMemory()))
                 .add(Network.createLayer("3", p)
                     .add(new SpatialPooler())
                     .using(l2.getConnections()))
@@ -1698,7 +1697,7 @@ public class PersistenceAPITest extends ObservableTestBase {
             .add(Network.createRegion("r2")
                 .add(Network.createLayer("1", p)
                     .alterParameter(KEY.AUTO_CLASSIFY, Boolean.TRUE)
-                    .add(new OldTemporalMemory())
+                    .add(new TemporalMemory())
                     .add(new SpatialPooler())
                     .add(Sensor.create(FileSensor::create, SensorParams.create(
                         Keys::path, "", ResourceLocator.path("days-of-week-stream.csv"))))))
@@ -1723,7 +1722,7 @@ public class PersistenceAPITest extends ObservableTestBase {
             .add(Network.createLayer("1", p)
                 .alterParameter(KEY.AUTO_CLASSIFY, true)
                 .add(Anomaly.create())
-                .add(new OldTemporalMemory())
+                .add(new TemporalMemory())
                 .add(new SpatialPooler())
                 .add(sensor)));
 
@@ -1739,14 +1738,14 @@ public class PersistenceAPITest extends ObservableTestBase {
             .add(Network.createRegion("r1")
                 .add(Network.createLayer("2", p)
                     .add(Anomaly.create())
-                    .add(new OldTemporalMemory()))
+                    .add(new TemporalMemory()))
                 .add(Network.createLayer("3", p)
                     .add(new SpatialPooler()))
                 .connect("2", "3"))
             .add(Network.createRegion("r2")
                 .add(Network.createLayer("1", p)
                     .alterParameter(KEY.AUTO_CLASSIFY, Boolean.TRUE)
-                    .add(new OldTemporalMemory())
+                    .add(new TemporalMemory())
                     .add(new SpatialPooler())
                     .add(Sensor.create(FileSensor::create, SensorParams.create(
                         Keys::path, "", ResourceLocator.path("rec-center-hourly.csv"))))))
@@ -1771,7 +1770,7 @@ public class PersistenceAPITest extends ObservableTestBase {
             .add(Network.createLayer("1", p)
                 .alterParameter(KEY.AUTO_CLASSIFY, true)
                 .add(Anomaly.create())
-                .add(new OldTemporalMemory())
+                .add(new TemporalMemory())
                 .add(new SpatialPooler())
                 .add(sensor)));
 
@@ -1787,7 +1786,7 @@ public class PersistenceAPITest extends ObservableTestBase {
             .add(Network.createLayer("1", p)
                 .alterParameter(KEY.AUTO_CLASSIFY, true)
                 .add(Anomaly.create())
-                .add(new OldTemporalMemory())
+                .add(new TemporalMemory())
                 .add(new SpatialPooler())));
                 
         return network;
@@ -1806,7 +1805,7 @@ public class PersistenceAPITest extends ObservableTestBase {
             .add(Network.createLayer("1", p)
                 .alterParameter(KEY.AUTO_CLASSIFY, true)
                 .add(Anomaly.create())
-                .add(new OldTemporalMemory())
+                .add(new TemporalMemory())
                 .add(new SpatialPooler())
                 .add(sensor)));
 
@@ -1924,7 +1923,7 @@ public class PersistenceAPITest extends ObservableTestBase {
         Network network = Network.create("test network", p)
             .add(Network.createRegion("r1")
                 .add(Network.createLayer("1", p)
-                    .add(new OldTemporalMemory())
+                    .add(new TemporalMemory())
                     .add(sensor)));
 
         network.start();
