+++ vocabulary.dictionary
@@ -233,3 +233,4 @@ col
 groupby
 itertools
 rawtypes
+squelched
